Fix warning in squeeze.h

+++ lotus/core/providers/cpu/tensor/squeeze.h
@@ -24,8 +24,8 @@ class SqueezeBase {
       std::vector<int64_t> axes) {
     int j = 0;
     std::vector<int64_t> output_shape;
-    for (int i = 0; i < input_shape.size(); ++i) {
-      if (j < axes.size() && axes[j] == i) {
+    for (size_t i = 0; i < input_shape.size(); ++i) {
+      if (j < axes.size() && axes[j] == static_cast<int64_t>(i)) {
         LOTUS_ENFORCE(input_shape[i] == 1, "Dimension of input ", i,
                       " must be 1 instead of ", input_shape[i]);
         ++j;
